requirements.txt
# === 向量化功能專用版本（推薦用於 Render 部署）===
# 只包含知識點自動關聯功能必需的套件

# === 核心 Web 框架與部署 ===
flask==2.3.3
gunicorn==21.2.0

# === 資料庫連接 ===
psycopg2-binary==2.9.5

# === 用戶認證與安全 ===
flask-jwt-extended==4.5.3
bcrypt==4.0.1

# === 環境變數管理 ===
python-dotenv==1.0.0

# === 知識點向量化功能（核心功能）===
numpy==2.0.2
sentence-transformers==2.2.2
scikit-learn==1.5.2

# === 可選功能 ===
# 如需要 AI 對話功能，取消註解以下行：
# openai==0.28.1
# google-generativeai==0.3.1
# requests==2.31.0